cast to float instead of int

+++ custom_components/daikin_altherma/number.py
@@ -95,7 +95,7 @@ class GenericOperationControl(NumberEntity, CoordinatorEntity):
         return 'box'
 
     async def async_set_native_value(self, value: float) -> None:
-        await self._api.device.climate_control.call_operation(self._operation, int(value), validate=False)
+        await self._api.device.climate_control.call_operation(self._operation, float(value), validate=False)
         await self.coordinator.async_request_refresh()
 
     @property
@@ -163,7 +163,7 @@ class AlthermaUnitTemperatureControl(NumberEntity, CoordinatorEntity):
 
     async def async_set_native_value(self, value: float) -> None:
         key, _ = self._get_value_config()
-        await self._api.device.climate_control.call_operation(key, int(value))
+        await self._api.device.climate_control.call_operation(key, float(value))
         await self.coordinator.async_request_refresh()
 
     @property
